Move fixtures to tests.core.conftest.py
